added main.cpp moved all under Emulator
