+++ Emulator/doc/SWDesign.plantuml
