+++ Emulator/doc/usecase.plantuml
