--- README.md
@@ -1,8 +1,0 @@
-# Demo
-
-I will uppload code later
-=======
-I will uppload code later
-
-Hi this is reza
-I'm adding another line here.
