--- a_new_file
@@ -1,1 +1,0 @@
-This is a file:
